+++ examples/apps/web-form/src/store/index.js
@@ -1,12 +1,11 @@
 import { createStore } from "@inglorious/web"
 
-import { app } from "../app"
 import { userForm } from "../user-form/user-form"
 import { entities } from "./entities"
 import { middlewares } from "./middlewares"
 
 export const store = createStore({
-  types: { app, userForm },
+  types: { userForm },
   entities,
   middlewares,
 })