+++ examples/apps/web-table/src/store/index.js
@@ -1,12 +1,11 @@
 import { createStore } from "@inglorious/web"
 
-import { app } from "../app"
 import { productTable } from "../product-table/product-table"
 import { entities } from "./entities"
 import { middlewares } from "./middlewares"
 
 export const store = createStore({
-  types: { app, productTable },
+  types: { productTable },
   entities,
   middlewares,
 })