+++ examples/apps/web-todomvc-cs/src/components/footer/handlers.js
@@ -1,0 +1,3 @@
+export function filterClick(entity, id) {
+  entity.activeFilter = id
+}