+++ examples/apps/web-todomvc-cs/src/components/footer/index.js
@@ -1,0 +1,4 @@
+import * as handlers from "./handlers"
+import { render } from "./render"
+
+export const footer = { ...handlers, render }