+++ examples/apps/web-todomvc-cs/src/components/form/handlers.js
@@ -1,0 +1,7 @@
+export function inputChange(entity, value) {
+  entity.value = value
+}
+
+export function formSubmit(entity) {
+  entity.value = ""
+}