+++ examples/apps/web-todomvc/src/store/index.js
@@ -1,6 +1,5 @@
 import { createStore } from "@inglorious/web"
 
-import { app } from "../app"
 import { footer } from "../footer/footer"
 import { form } from "../form/form"
 import { list } from "../list/list"
@@ -8,7 +7,7 @@ import { entities } from "./entities"
 import { middlewares } from "./middlewares"
 
 export const store = createStore({
-  types: { app, form, list, footer },
+  types: { form, list, footer },
   entities,
   middlewares,
 })