+++ examples/apps/react-todomvc/CHANGELOG.md
@@ -1,5 +1,12 @@
 # todomvc
 
+## 1.1.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/store@7.1.3
+
 ## 1.1.9
 
 ### Patch Changes