+++ examples/apps/web-todomvc/CHANGELOG.md
@@ -1,5 +1,13 @@
 # lit-todomvc
 
+## 0.1.14
+
+### Patch Changes
+
+- Add a client-server Inglorious Web demo
+- Updated dependencies
+  - @inglorious/web@2.3.0
+
 ## 0.1.13
 
 ### Patch Changes