+++ packages/renderer-react-dom/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inglorious/renderer-react-dom
 
+## 12.0.3
+
+### Patch Changes
+
+- @inglorious/engine@13.0.3
+
 ## 12.0.2
 
 ### Patch Changes